requirements.txt
python-telegram-bot==21.5
openai>=1.52.0